pyproject.toml
[project]
description = "Project GHOST SHIP - Core Engine (Stateless Data Normalization)"
name = "gs-mcp-core"
readme = "README.md"
requires-python = ">=3.12"
version = "3.0.0"

dependencies = [
  "fastapi[standard]>=0.115.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.4.0",
  "google-genai>=1.0.0",
  "tavily-python>=0.7.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
